refactor(server): extract embedded frontend renderer
